Highest and Lowest Test Scores
Hi Zach, this is an example I created using ArrayList (I hope your instructor is fine with you using this), and using the max, min classes below. I like the ArrayList because its size is super flexible and you can add to it as much as you want. Hope this makes sense. And I left out the average - pretty sure you can do that on your own. Good luck!